BT Polling enabled
Setup a 3s check to ensure BT is still connected. If BT is lost, the status bar will turn red. If connected, the status bar will turn Azure blue to match that of the title bar area. May change this to green but hesitated now to avoid "Christmas" colors. Orange would be. a good alternative color as it is complementary to the azure blue.